Create an ASTUtil module and move namedVars to it from the typechecker.
Going to want it elsewhere too.
