+++ saw-script.cabal
@@ -106,6 +106,7 @@ library
 
   exposed-modules:
     SAWScript.AST
+    SAWScript.ASTUtil
     SAWScript.AutoMatch
     SAWScript.AutoMatch.Interaction
     SAWScript.AutoMatch.ArgMapping
